refactor #4: add validation user for delete event

+++ src/main/java/cz/cvut/iarylser/facade/EventFacadeImpl.java
@@ -1,6 +1,5 @@
 package cz.cvut.iarylser.facade;
 
-import cz.cvut.iarylser.controller.EventController;
 import cz.cvut.iarylser.dao.DTO.EventDTO;
 import cz.cvut.iarylser.dao.DTO.PurchaseRequest;
 import cz.cvut.iarylser.dao.DTO.TicketDTO;
@@ -9,13 +8,9 @@ import cz.cvut.iarylser.dao.entity.User;
 import cz.cvut.iarylser.dao.mappersDTO.EventMapperDTO;
 import cz.cvut.iarylser.service.AuthService;
 import cz.cvut.iarylser.service.EventService;
-import cz.cvut.iarylser.service.EventServiceImpl;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -32,7 +27,7 @@ public class EventFacadeImpl implements EventFacade {
     @Override
     public EventDTO create(EventDTO dto) {
         Event entity = eventMapper.toEntity(dto);
-        User currentUser = authService.authenticationUser();
+        User currentUser = authService.getUser();
         entity.setOrganizer(currentUser.getUsername());
         Event result = eventService.create(entity);
         return eventMapper.toDTO(result);
@@ -40,7 +35,7 @@ public class EventFacadeImpl implements EventFacade {
 
     @Override
     public EventDTO update(Long id, EventDTO dto) throws EntityNotFoundException, IllegalStateException, IllegalAccessException {
-        User currentUser = authService.authenticationUser();
+        User currentUser = authService.getUser();
         if(!Objects.equals(dto.getOrganizer(), currentUser.getUsername())) {
             log.warn("User {} is not the same as the organizer.", currentUser.getUsername());
             throw new IllegalAccessException("User" + currentUser.getUsername() + " is not the same as the organizer.");
@@ -51,7 +46,13 @@ public class EventFacadeImpl implements EventFacade {
     }
 
     @Override
-    public boolean delete(Long id) {
+    public boolean delete(Long id) throws IllegalAccessException {
+        User currentUser = authService.getUser();
+        Event eventToDelete = eventService.getById(id);
+        if(!Objects.equals(currentUser.getUsername(), eventToDelete.getOrganizer())) {
+            log.warn("User {} is not the same as the organizer.", currentUser.getUsername());
+            throw new IllegalAccessException("User" + currentUser.getUsername() + " is not the same as the organizer.");
+        }
         return eventService.delete(id);
     }
 
@@ -75,7 +76,7 @@ public class EventFacadeImpl implements EventFacade {
 
     @Override
     public List<TicketDTO> purchaseTicket(Long eventId, PurchaseRequest request) throws EntityNotFoundException, IllegalStateException {
-        User currentUser = authService.authenticationUser();
+        User currentUser = authService.getUser();
         request.setCustomer(currentUser.getUsername());
 
         return eventService.purchaseTicket(eventId,request);
@@ -83,14 +84,14 @@ public class EventFacadeImpl implements EventFacade {
 
     @Override
     public boolean like(Long eventId, Long userId) {
-        User currentUser = authService.authenticationUser();
+        User currentUser = authService.getUser();
 
         return eventService.like(eventId, currentUser.getId());
     }
 
     @Override
     public boolean unlike(Long eventId, Long userId) { // todo fix userId
-        User currentUser = authService.authenticationUser();
+        User currentUser = authService.getUser();
         return eventService.unlike(eventId,currentUser.getId());
     }
